PHP Finished and all connections to local host done
So all the code on my side is finished and everything is connected to the local server I just need to check and make sure it all reads and writes properly from and to the database, as long as it does we are good to go. Also check the index.html I've left a message for you there.